move type to servicecfg
