+++ lib/service/servicecfg/externalauditstorage.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package config
+package servicecfg
 
 // ExternalAuditStorageConfiguration contains the arguments to configure the
 // External Audit Storage.
